app/requirements.txt
asgiref==3.7.2
boto3==1.26.102
Django==4.2.8
django-cognito-jwt==0.0.4
django-cors-headers==4.3.1
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
pytz==2023.3.post1
requests==2.28.2
sqlparse==0.4.4
stripe==5.4.0
typing-extensions==4.8.0
whitenoise==6.6.0